+++ evmlab/opcodes.py
@@ -2,55 +2,8 @@ from copy import copy
 import collections
 import binascii
 
-try:
-    from ethereum.utils import parse_int_or_hex,decode_hex,remove_0x_head,bytearray_to_bytestr,encode_hex
-except:
+from . import parse_int_or_hex,decode_hex,remove_0x_head,bytearray_to_bytestr,encode_hex
 
-    print("ethereum.utils not available, using fallback methods")
-
-    def encode_hex(b):
-        if isinstance(b, str):
-            b = bytes(b, 'utf-8')
-        if isinstance(b, (bytes, bytearray)):
-            return str(binascii.hexlify(b), 'utf-8')
-        raise TypeError('Value must be an instance of str or bytes')
-
-    def to_string(value):
-        if isinstance(value, bytes):
-            return value
-        if isinstance(value, str):
-            return bytes(value, 'utf-8')
-        if isinstance(value, int):
-            return bytes(str(value), 'utf-8')
-
-    def bytearray_to_bytestr(value):
-        return bytes(value)
-
-    def decode_hex(s):
-        if isinstance(s, str):
-            return bytes.fromhex(s)
-        if isinstance(s, (bytes, bytearray)):
-            import binascii
-            return binascii.unhexlify(s)
-        raise TypeError('Value must be an instance of str or bytes')
-
-    def is_numeric(x): return isinstance(x, int)
-
-    def remove_0x_head(s):
-        return s[2:] if s[:2] in (b'0x', '0x') else s
-
-    def big_endian_to_int(value):
-        return int.from_bytes(value, byteorder='big')
-
-    def parse_int_or_hex(s):
-        if is_numeric(s):
-            return s
-        elif s[:2] in (b'0x', '0x'):
-            s = to_string(s)
-            tail = (b'0' if len(s) % 2 else b'') + s[2:]
-            return big_endian_to_int(decode_hex(tail))
-        else:
-            return int(s)
 # Taken from https://github.com/ethereum/pyethereum/blob/develop/ethereum/opcodes.py
 # Done this way to reduce dependencies a bit
 # schema: [opcode, ins, outs, gas]
